third_party/WebKit/LayoutTests/fast/html/imports/import-events-inline.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
var list = [];
function handle(event)
{
    list.push(event.type + '@' + event.target.id);
    if (list.length < 3)
        return;
    expected = ['load@shouldLoad', 'error@shouldError', 'click@toClick'];
    shouldBe('window.list', 'expected');
    finishJSTest();
}
</script>
<link id="harness" rel="import" href="resources/events-inline.html">
</head>
<body>
<script>
window.jsTestIsAsync = true;
description('Test that inline event handler is fired on an imported document.');
</script>
</body>
</html>
