third_party/WebKit/LayoutTests/fast/html/imports/import-events-inline-expected.txt

Test that inline event handler is fired on an imported document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.list is expected
PASS successfullyParsed is true

TEST COMPLETE
